+++ NEWS.md
@@ -1,3 +1,6 @@
+# Development version (0.2.1.9000)
+* Adds `.SOURCE("table_name")` and `mark_source(obj)` to add variable (and table) lineage to non-blueprint-created tables. Particularly useful for including data fetched from the web into lineage
+
 # blueprintr 0.2.1
 * Implements variable lineage through `options(blueprintr.use_variable_uuids = TRUE)`. Visualize variable lineage with `blueprintr::vis_variable_lineage` or inspect specific parts of the graph with an igraph generated with `blueprintr::load_variable_lineage`.
 * Visualize blueprint table lineage with `blueprintr::vis_table_lineage` and inspect specific parts of the graph with `blueprintr::load_table_lineage`.